requirements.txt
blinker==1.9.0
click==8.2.1
fire==0.7.1
Flask==3.1.1
fonttools==4.59.1
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.0.0
MarkupSafe==3.0.2
numpy==2.2.6
opencv-python-headless==4.12.0.88
pdf2docx==0.5.8
PyMuPDF==1.26.3
python-docx==1.2.0
termcolor==3.1.0
typing_extensions==4.14.1
Werkzeug==3.1.3
